Add text orientation support
With this commit, TextInfo will now also contain rotation information.
I.e. if there is a rotated piece of text on the page and the OCR tool
manages to correctly detect that, this information will now be used to
write the text correctly in the output PDF.

Tesseract supports such detection in the "(1) Automatic page segmentation
with OSD." mode. In that case in the "title" attribute there is a
"textangle" key, which has the degrees of rotation.

This will, most likely, require some tweaks in the algorithms of merging
words into lines.

I wasn't able to verify whether it affected existing tests, as I have 13
failing test before and after the commit.

The functionality was tested manually. I would have added some tests,
but since tests already fail on my machine in comparisons, they would
need to be remade anyway...