Add support for more image to ML input conversion options
A lot more padding options are now supported. The ones we had before
were extracted as BOTTOM_RIGHT_BLACK and SYMMETRIC_BLACK. At the same
time a bunch of new ones were added.

EDGE padding strategy was encountered in EasyOCR. In some of their
models the extend the edges of the image. This is pretty impactful on
the results, as with just black padding, EasyOCR models often produce
pipes/brackets/... at the padded borders.

GRAY padding is, somewhat unexpectedly, required for the PaddleOCR
models to work properly with our code... Usually you would pad the
image _before_ doing value normalization for the input tensors. But in
PaddleOCR they resize the image, normalize it and only then they add 0
padding. So for it to work properly we need a value, that will become 0
after normalization. Since they use 0.5 for mean and std, just using 128
gray is close enough for it to start working as expected.

I haven't encountered WHITE padding in the wild yet, but I might as well
add it anyway.

Another major generalization was the addition of different channel
configurations for input tensor buffers. In addition to RGB we now
support BGR and Grayscale. BGR appears relatively frequently in ML
models as it is the default configuration with OpenCV. PaddleOCR models
are one of the examples. Grayscale was encountered in EasyOCR text
recognition models and seems like a reasonable option to have. The
configuration itself is encoded in the ImageChannelConfiguration enum.

Support for non-fixed size model inputs has also been added. In OnnxTR
models had a static width and height, but a lot of other models have
them as variables. In such cases there is usually some logic in the
library code itself on what the optimal range is. So now there is a
minimum and maximum width/height, in which the resized image should
fit.

Another, somewhat unexpected, option is width/height multiples. In some
models, like the PaddleOCR text detection model, the convolution
operations are rather strict, so the input image dimensions should be
a multiple of some integer value. For simplicity's sake, dimensions are
just bumped-up to the next multiple value and max width/height values
are checked in the constructor to be valid regards to the multiple.

Since with these changes each image in the batch can have different
"target" dimensions based on the new resizing algorithm, I had to alter
batching processing a bit. First we now calculate "target" dimensions
for each image separately, then calculate the final target dimensions
for the batch to fit them all. And only after that we do image resizing
operations. Before we could do resizing immediately without a
pre-calculation step, as we were getting a fixed batch size from the
input parameters from the very start.

Since there are a lot of options now, they were extracted from
OnnxInputProperties into a separate ImageResizeOptions class. The
original OnnxInputProperties constructor has been deprecated as a
result, but should still be working fine.