Add StringMapper as a Vocabulary alternative
In some of the more general PaddleOCR recognition models, the index
mapping has cases, which are not covered by Vocabulary. Some of the
have code points outside the BMP range, and they also sometimes include
Unicode flag sequences, which are several code points. It is much less
memory efficient, so it still makes sense to use Vocabulary, if
possible.