Improve orientation predictor results for long words/lines
While trying to get EasyOCR and PaddleOCR text detection models to work,
a problem popped-up with the OnnxTR orientation predictor model. It
works fine for words, and since OnnxTR text detection step outputs
words, it works fine.

But in EasyOCR and PaddleOCR the output is lines, which by their nature
can be very long. And when long lines get squeezed to fit a 256x256
square with a ratio-preserving resize, too much information is lost in
one of the dimensions, which causes the output to be pretty random.

As a workaround, now we will truncate input images, when the ratio is
too extreme. On the surface, this shouldn't cause any issues, as even
if you truncate a perfectly fine word, it would still look like text,
so there will still be enough information to work with.

The truncation is done symmetrically on both sides, which is pretty
important. This way we have a higher chance to catch text and not the
margins/padding. Doing it only on one side gives worse results.

Somewhat surprisingly, based on the tests that required updating, it
also improved results there with the docTR model, which is also a
positive sign.