Add CtcLabelPostProcessor recognition post-processor
This post-processor contains the logic, which is used by both EasyOCR
(CTCLabelConverter) and PaddleOCR (CTCLabelDecode). It is, pretty much,
the same as CrnnPostProcessor, which we already have, but with some
minor differences:

* Blank character is the first one instead of being the last one.

* It uses a IOutputLabelMapper<String> instead of a Vocabulary. This
  makes it less efficient, but we cannot use Vocabulary for all
  PaddleOCR models, as in some of the models labels do not map cleanly
  to char.

So CrnnPostProcessor was abstracted away into BasicLabelPostProcessor,
and it looks like it will be applicable for a lot of models...

Additionally, while we currently make sure, that the output probs
vector has the same size as labelDimension() on the shape validation
step, it would be better to not rely on that to make the code a bit
more robust. So the post-processor logic was changed appropriately.