Make text box splitting before text recognition optional
In OnnxTR, since the model input size is static, there is logic to
split long text boxes into smaller ones and to then merge the results
back, so that they match the original images.

This is not great for models, which work on lines... The text boxes
there are long, but it is fine for those models, as they need more info
and they process spaces. So for such models, if there is text box
splitting, the end results starts missing spaces and there are repeating
parts of words.

This commit makes text box splitting an additional option within
OnnxRecognitionPredictorProperties. For backward-compatibility text box
splitting is still the default behavior.