Fix channel order handling in BufferedImageUtil.toBchwInput
Originally, when I was writing this code, I was under the assumption,
that `Raster` is a pretty low-level interface. So I blindly assumed,
that if pixel values are in a BGR order in the data buffer, then
calling `Raster.getSample` with band 0 would return the blue value.

But that is not the case. The bands are RGB ordered, and the sampling
model does the index conversion for us. So we were passing the wrong
values, which causes the red and blue values being swapped...

Since most tests involved black text on white backgrounds, there are
not a lot of meaningful changes in the results. In some cases the
result is a bit better, in some it is a bit worse.

Additionally added a basic test to cover this in the future.

This issue is present only in the Java version of the library.