Add PaddleOCR predictor creation methods
In PaddleOCR models come together with YAML config files, which include
information on how to set up pre-processors and post-processors. In
practice, a lot more configuration comes from the hardcoded values and
the general OCR configuration file, but still.

So instead of hardcoding parameters, like we did with OnnxTR, it makes
more sense to just parse the config files ourselves. Since models are
distributed in archives, it has a predictable structure, so you could
just pass the path to the model directory.

Parsing of the config file into data structures is done in code. This
should, in theory, simplify porting. Additionally, when I tried doing
it "properly" via Jackson, PaddleOCR config files weren't mapping
nicely, and I was forced to add custom deserialization logic anyway. At
least this way it is easier to reason about.