Decrease default OnnxTR batch sizes
At the moment we only run the models on the CPU, in which case the
batch size has almost no effect on performance, but still has a major
effect on memory usage. Decreased the default batch sizes in OnnxTR for
now to limit possible OOM cases.

Since OnnxTR has static dimensions on model inputs, this has no effect
on the results.