Add EasyOCR text recognition predictor creation methods
All the models have the same input properties, only the label mapping
is different. So I've made a single method for all the EasyOCR models,
with model path + mapper. Label mappers are static in the same way we
made it for OnnxTR models.

Initially I had EasyOcrMapper as just a static class for storing string
mappers, but it is somewhat convenient for it to be a strongly-typed
alias to StringMapper to make the API a bit more intuitive.