Add post-processor for EasyOCR text detection models
The post-processing there is, for the most part, also similar to what
we have now, with a couple of exceptions:

* The model returns two channels instead of one. The first one detect
  the text glyphs themselves. While the seconds one outputs
  "connections" between the glyphs.
* The mask for finding text contours is based on the sum of both
  channels instead of just the first, text prediction, one.
* Text score calculation is based on `max` instead of `mean`.
* Text box enlargement is calculated differently.
* Text boxes are merged into lines at the end to pass to recognition.

Since you can use the PaddleOCR detection model to get lines, I am not
quite sure, if there is a practical need for this. But it works. Maybe
it could be useful for some of the more esoteric languages.